motion_cam_requirements.txt
RPi.GPIO
requests
python-socketio
